+++ docs/changes.md
@@ -1,7 +1,8 @@
 # 最新动态
 
 2021/04/06
-  * add idle\_queue\_ex/timer\_queue\_ex for script binding.
+  * 增加 idle\_queue\_ex/timer\_queue\_ex 方便脚本绑定时使用。
+  * 增加 hscroll\_label检测焦点状态下长度变化重新播放动画的机制（感谢智明提供补丁）
 
 2021/04/04
   * 增加 gui\_app\_start\_ex，方便嵌入式系统指定资源路径。